requirements.txt
requests==2.26.0
fastapi==0.68.0
uvicorn==0.15.0
python-dotenv==0.19.0
aiofiles==0.7.0
python-multipart==0.0.5
jinja2==3.0.1
Markdown==3.3.4
pytest==6.2.4
